Unlock Cut Cars Script
Drop the .bat and .py in the folder of your save for ITC or SBX, run the .bat and follow the prompts to unlock the cut cars as buyable options in the shop. This does not give you the cars for free in your garage and only adds the cars as buyable options. You might need to rerun the script on your save after unlocking the Devil Z and/or Skull Bullet.