Update save_convert.py
Fixed the bug with the controller rebinds breaking on converted saves. It will now use the default controls of the appropriate game so make sure to rebind your controls after a save conversion similar to how you need to reset your license plate. If you have an old converted save with the broken controller bug simply just double convert the save back and forth once more with the script.